+++ packages/apollo-react-native/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [2.0.2-alpha.0](https://github.com/uplift-ltd/nexus/compare/@uplift-ltd/apollo-react-native@2.0.1...@uplift-ltd/apollo-react-native@2.0.2-alpha.0) (2022-04-04)
+
+**Note:** Version bump only for package @uplift-ltd/apollo-react-native
+
+
+
+
+
 ### [2.0.1](https://github.com/uplift-ltd/nexus/compare/@uplift-ltd/apollo-react-native@2.0.0...@uplift-ltd/apollo-react-native@2.0.1) (2022-02-02)
 
 **Note:** Version bump only for package @uplift-ltd/apollo-react-native