+++ packages/file-uploads/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.5.1-alpha.0](https://github.com/uplift-ltd/nexus/compare/@uplift-ltd/file-uploads@1.5.0...@uplift-ltd/file-uploads@1.5.1-alpha.0) (2022-04-04)
+
+**Note:** Version bump only for package @uplift-ltd/file-uploads
+
+
+
+
+
 ## [1.5.0](https://github.com/uplift-ltd/nexus/compare/@uplift-ltd/file-uploads@1.4.6...@uplift-ltd/file-uploads@1.5.0) (2022-03-04)